Move `Column` into separate package
